+++ pkg/barback/lib/src/barback.dart
@@ -110,9 +110,8 @@ class Barback {
   /// To the extent that [transformers] is similar to the previous transformer
   /// phases for [package], the existing asset graph will be preserved.
   ///
-  /// Elements of the inner iterable of [transformers] must be either
-  /// [Transformer]s or [TransformerGroup]s.
-  void updateTransformers(String package,
-          Iterable<Iterable> transformers) =>
+  /// Elements of the inner iterable of [transformers] must be [Transformer]s,
+  /// [TransformerGroup]s, or [AggregateTransformer]s.
+  void updateTransformers(String package, Iterable<Iterable> transformers) =>
       _graph.updateTransformers(package, transformers);
 }