+++ pkg/barback/test/transformer/lazy_assets.dart
@@ -11,6 +11,7 @@ import 'declare_assets.dart';
 /// Like [DeclareAssetsTransformer], but lazy.
 class LazyAssetsTransformer extends DeclareAssetsTransformer
     implements LazyTransformer {
-  LazyAssetsTransformer(Iterable<String> declared, [Iterable<String> emitted])
-      : super(declared, emitted);
+  LazyAssetsTransformer(Iterable<String> declared, {Iterable<String> emitted,
+        String input})
+      : super(declared, emitted: emitted, input: input);
 }